# src/main/resources/swagger_api_doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Airport",
      "description": "Airport CRUD APIs"
    },
    {
      "name": "Flight",
      "description": "Flight CRUD APIs"
    },
    {
      "name": "Search",
      "description": "Search APIs"
    }
  ],
  "paths": {
    "/flights/{id}": {
      "get": {
        "tags": [
          "Flight"
        ],
        "operationId": "getFlightById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Flight"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Flight"
        ],
        "operationId": "updateFlight",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flight"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Flight"
        ],
        "operationId": "deleteFlight",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/airports/{id}": {
      "get": {
        "tags": [
          "Airport"
        ],
        "operationId": "getAirportById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Airport"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Airport"
        ],
        "operationId": "updateAirport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Airport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Airport"
        ],
        "operationId": "deleteAirport",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/flights": {
      "post": {
        "tags": [
          "Flight"
        ],
        "operationId": "createFlight",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Flight"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/airports": {
      "post": {
        "tags": [
          "Airport"
        ],
        "operationId": "createAirport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Airport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/flights/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "operationId": "searchFlights",
        "parameters": [
          {
            "name": "departureCity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "arrivalCity",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "departureDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "returnDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Airport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "city": {
            "type": "string"
          }
        }
      },
      "Flight": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "departureAirport": {
            "$ref": "#/components/schemas/Airport"
          },
          "arrivalAirport": {
            "$ref": "#/components/schemas/Airport"
          },
          "departureDateTime": {
            "type": "string"
          },
          "returnDateTime": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      }
    }
  }
}